requirements.txt
nose==1.3.3
python-memcached==1.53
